[Improvement-3369][api] Introduce alert group and users service interface for clear code (#4758)
