+++ dolphinscheduler-api/src/main/java/org/apache/dolphinscheduler/api/service/BaseService.java
@@ -16,19 +16,16 @@
  */
 package org.apache.dolphinscheduler.api.service;
 
-import java.text.MessageFormat;
-import java.util.Map;
-
-import javax.servlet.http.Cookie;
-import javax.servlet.http.HttpServletRequest;
-
 import org.apache.dolphinscheduler.api.enums.Status;
 import org.apache.dolphinscheduler.api.utils.Result;
 import org.apache.dolphinscheduler.common.Constants;
 import org.apache.dolphinscheduler.common.enums.UserType;
 import org.apache.dolphinscheduler.common.utils.HadoopUtils;
-import org.apache.dolphinscheduler.common.utils.StringUtils;
 import org.apache.dolphinscheduler.dao.entity.User;
+
+import java.io.IOException;
+import java.text.MessageFormat;
+import java.util.Map;
 
 /**
  * base service
@@ -117,9 +114,9 @@ public class BaseService {
      * create tenant dir if not exists
      *
      * @param tenantCode tenant code
-     * @throws Exception if hdfs operation exception
+     * @throws IOException if hdfs operation exception
      */
-    protected void createTenantDirIfNotExists(String tenantCode) throws Exception {
+    protected void createTenantDirIfNotExists(String tenantCode) throws IOException {
 
         String resourcePath = HadoopUtils.getHdfsResDir(tenantCode);
         String udfsPath = HadoopUtils.getHdfsUdfDir(tenantCode);
